// src/Modules/Dynamic/Admin_1_mini_1_main/Models/IPlanCategory.ts
export interface IPlanCategory {
  categoryId: number;
  categoryName: string;
  activeSwitch?: string | null;
  cerateDate?: string;
  updateDate?: string;
  createdBy?: string;
  updateBy?: string;
}

// categoryId	[...]
// categoryName	[...]
// activeSwitch	[...]
// cerateDate	[...]
// updateDate	[...]
// createdBy	[...]
// updateBy	[...]